modify repositories for socket
