+++ src/server/AServer.class.cpp
@@ -1,25 +1,15 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   main.hpp                                           :+:      :+:    :+:   */
+/*   AServer.class.cpp                                  :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: tlafont <tlafont@student.42angouleme.      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/02/13 10:11:36 by tlafont           #+#    #+#             */
-/*   Updated: 2023/02/14 08:27:51 by tlafont          ###   ########.fr       */
+/*   Created: 2023/02/14 08:29:43 by tlafont           #+#    #+#             */
+/*   Updated: 2023/02/14 08:30:25 by tlafont          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#ifndef MAIN_HPP
-# define MAIN_HPP
+#include "../../inc/server/AServer.class.hpp"
 
-// includes
-#include <iostream>
 
-// sockets
-#include "sockets/ISocket.class.hpp"
-#include "sockets/BindSocket.class.hpp"
-#include "sockets/ClientSocket.class.hpp"
-#include "sockets/ListenSocket.class.hpp"
-
-#endif
