+++ src/sockets/BindSocket.class.cpp
@@ -1,16 +1,16 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ClientSocket.class.cpp                             :+:      :+:    :+:   */
+/*   BindSocket.class.cpp                               :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: tlafont <tlafont@student.42angouleme.      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/10 14:30:23 by tlafont           #+#    #+#             */
-/*   Updated: 2023/02/13 11:43:56 by tlafont          ###   ########.fr       */
+/*   Updated: 2023/02/13 11:17:33 by tlafont          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../inc/ClientSocket.class.hpp"
+#include "../../inc/sockets/BindSocket.class.hpp"
 
 /*
 *  @brief   Overload constructor.
@@ -18,22 +18,21 @@
 *  @param   int, int, int, int , unsigned long
 *  @return  void
 */
-ClientSocket::ClientSocket(int dom, int serv, int protoc, int port, u_long interf):
+BindSocket::BindSocket(int dom, int serv, int protoc, int port, u_long interf):
 ISocket(dom, serv, protoc, port, interf)
 {
 	//establish the connection to network and test
 	this->_connec = connectToNetwork(this->_sock_fd, this->_addr);
-	std::cout << _connec << "\n";
 	testConnection(this->_connec);
 }
 
 /*
 *  @brief   Assignment operator.
 *           copy other socket in this socket
-*  @param   ClientSocket &
-*  @return  ClientSocket &
+*  @param   BindSocket &
+*  @return  BindSocket &
 */
-ClientSocket	&ClientSocket::operator=(ClientSocket const &rhs)
+BindSocket	&BindSocket::operator=(BindSocket const &rhs)
 {
 	(void)rhs;
 	return (*this);
@@ -45,9 +44,9 @@ ClientSocket	&ClientSocket::operator=(ClientSocket const &rhs)
 *  @param   int , struct sockaddr_in
 *  @return  int
 */
-int ClientSocket::connectToNetwork(int sock, struct sockaddr_in addr) const
+int BindSocket::connectToNetwork(int sock, struct sockaddr_in addr) const
 {
-	return (connect(sock, (struct sockaddr *)&addr, sizeof(addr)));
+	return (bind(sock, (struct sockaddr *)&addr, sizeof(addr)));
 }
 
 /*
@@ -56,6 +55,6 @@ int ClientSocket::connectToNetwork(int sock, struct sockaddr_in addr) const
 *  @param   void
 *  @return  void
 */
-ClientSocket::~ClientSocket()
+BindSocket::~BindSocket()
 {
 }
